refactor package structure to use directory for config routines
