+++ fv3config/data/__init__.py
@@ -1,0 +1,3 @@
+import os
+
+DATA_DIR = os.path.dirname(os.path.abspath(__file__))
