+++ tests/test_tables.py
@@ -5,11 +5,10 @@ from fv3config import (
     get_default_config, ConfigError,
 )
 from fv3config._tables import (
-    get_field_table_filename, get_diag_table_filename, get_data_table_filename,
-    get_microphysics_name_from_config, get_current_date_from_coupler_res,
-    get_current_date_from_config, update_diag_table_for_config
+    update_diag_table_for_config
 )
-
+from fv3config.config.derive import get_current_date, get_microphysics_name, \
+    get_field_table_filename, get_diag_table_filename, get_data_table_filename, _get_current_date_from_coupler_res
 
 test_directory = os.path.dirname(os.path.realpath(__file__))
 
@@ -110,7 +109,7 @@ class TableTests(unittest.TestCase):
         config = get_default_config()
         config['namelist']['gfs_physics_nml']['imp_physics'] = -1
         with self.assertRaises(NotImplementedError):
-            get_microphysics_name_from_config(config)
+            get_microphysics_name(config)
 
     def test_get_bad_diag_table_filename(self):
         diag_table_filename = '/not/a/path/diag_table'
@@ -139,7 +138,7 @@ class TableTests(unittest.TestCase):
         coupler_res_filename = os.path.join(rundir, 'coupler.res')
         with open(coupler_res_filename, 'w') as f:
             f.write(valid_coupler_res)
-        current_date = get_current_date_from_coupler_res(coupler_res_filename)
+        current_date = _get_current_date_from_coupler_res(coupler_res_filename)
         self.assertEqual(current_date, valid_current_date)
 
     def test_get_current_date_from_bad_coupler_res(self):
@@ -148,14 +147,14 @@ class TableTests(unittest.TestCase):
         with open(coupler_res_filename, 'w') as f:
             f.write(bad_coupler_res)
         with self.assertRaises(ConfigError):
-            get_current_date_from_coupler_res(coupler_res_filename)
+            _get_current_date_from_coupler_res(coupler_res_filename)
 
     def test_get_current_date_from_config(self):
         rundir = self.make_run_directory('test_rundir')
         config = get_default_config()
         config['namelist']['coupler_nml']['force_date_from_namelist'] = True
         config['namelist']['coupler_nml']['current_date'] = valid_current_date
-        current_date = get_current_date_from_config(config, os.path.join(rundir, 'INPUT'))
+        current_date = get_current_date(config, os.path.join(rundir, 'INPUT'))
         self.assertEqual(current_date, valid_current_date)
 
     def test_update_diag_table_for_config(self):
@@ -164,7 +163,7 @@ class TableTests(unittest.TestCase):
         diag_table_filename = os.path.join(rundir, 'diag_table')
         with open(diag_table_filename, 'w') as f:
             f.write(diag_table_test_in)
-        current_date = get_current_date_from_config(config_for_update_diag_table_test, input_directory)
+        current_date = get_current_date(config_for_update_diag_table_test, input_directory)
         update_diag_table_for_config(config_for_update_diag_table_test, current_date, diag_table_filename)
         with open(diag_table_filename) as f:
             self.assertEqual(diag_table_test_out, f.read())
